+++ cf/read_write/um/umread/c-lib/make.sh
@@ -1,2 +1,0 @@
-gcc -g -c -fPIC umfile_test.c
-ld -share -o umfile_test.so umfile_test.o